Move CellFaceType, DataContainer, ClusterStatistics from RigWellTargetMapping to RigWellTargetMappingTools
These types are not referenced in any RigWellTargetMapping public function
signature, so they belong in the tools class alongside the functions that
use them. RigWellTargetMapping now only owns the types that appear in its
public API (VolumeType, VolumeResultType, VolumesType, ClusteringLimits).
